requirements.txt
# V2.3 空间编码器依赖

# 核心依赖
torch>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
scikit-learn>=1.3.0

# API服务
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0

# 向量索引 (可选，推荐安装)
faiss-cpu>=1.7.4
# 或者使用GPU版本: faiss-gpu>=1.7.4

# 可视化
matplotlib>=3.7.0
